Remove LIONLoss stuff
